Separate the reusable layers
