+++ pytorch_forecasting/layers/_convolution/__init__.py
@@ -1,0 +1,3 @@
+from pytorch_forecasting.layers._convolution._reparam_large_kernel_conv import ReparamLargeKernelConv
+
+__all__ = ["ReparamLargeKernelConv"]
