+++ pytorch_forecasting/layers/_head/__init__.py
@@ -1,0 +1,3 @@
+from pytorch_forecasting.layers._head._flatten_head import Flatten_Head
+
+__all__ = ["Flatten_Head"]
